feat: inclusão de auto. usuário

+++ controller/ControllerBook.js
@@ -1,4 +1,6 @@
 const TaskBook = require("../models/Book");
+const mongoose = require('mongoose')
+const jwt = require("jsonwebtoken");
 
 class BookController {
   async index(req, res) {
@@ -13,8 +15,9 @@ class BookController {
 
   async store(req, res) {
     try {
-      // const [, token] = req.headers.authorization.split(" ");
-      // const { id: user_id } = jwt.decode(token);
+      const [, token] = req.headers.authorization.split(" ");
+      const { id: user_id } = jwt.decode(token);
+
 
       //VALIDAÇÕES
       const { titulo, autor, ISBN } = req.body
@@ -33,6 +36,7 @@ class BookController {
 
       // CRIAÇÃO BOOK
       const taskBook = await TaskBook.create({
+        user_id: user_id,
         titulo: req.body.titulo,
         autor: req.body.autor,
         ISBN: req.body.ISBN,
@@ -47,9 +51,14 @@ class BookController {
   }
   async update(req, res) {
     try {
-      const { _id } = req.params;
+      const [, token] = req.headers.authorization.split(" ");
+      const { id: user_id } = jwt.decode(token);
+      const { _id } = req.params
+      const { titulo, autor, ISBN } = req.body
 
-      // VALIDAÇÕES
+
+      // VALIDAÇÕES DE PREENCHIMENTO
+
       if(!titulo || titulo.length < 3 || titulo === ""){
         return res.status(400).json({msg:"Necessário preenchimento do titulo"})
       }
@@ -61,7 +70,7 @@ class BookController {
       if(!ISBN || ISBN.length < 7 || ISBN === ""){
         return res.status(400).json({msg:"ISBN mínimo 7 caracteres"})
       }     
-      
+
       //UPDATE DO BOOK
       const taskBook = {
         titulo: req.body.titulo,
@@ -69,7 +78,18 @@ class BookController {
         ISBN: req.body.ISBN,
       };
 
+      //VALIDAÇÃO DE ID
+      if (!mongoose.Types.ObjectId.isValid(_id)) {
+        return res.status(404).json({ msg: "book id inválido" });
+      }
+
+      const UserUpdateBook = await TaskBook.findById(_id, user_id)
+      if(!UserUpdateBook){
+      return res.status(403).json({ erro: "Voce nao pode alterar esse ebook" });         
+      }
+
       const updateBook = await TaskBook.findByIdAndUpdate(_id, taskBook);
+
 
       return res.status(200).json(updateBook);
     } catch (error) {
@@ -78,7 +98,15 @@ class BookController {
   }
   async destroy(req, res) {
     try {
+      const [, token] = req.headers.authorization.split(" ");
+      const { id: user_id } = jwt.decode(token);
       const { _id } = req.params;
+
+      //VALIDAÇÃO DE USUÁRIO AUTORIZADO
+      const UserDestroyBook = await TaskBook.findById(_id, user_id)
+      if(!UserDestroyBook){
+        return res.status(400).json({msg:"Usuário não autorizado"})
+      }
 
       const destroyBook = await TaskBook.findByIdAndDelete(_id);
 
@@ -89,6 +117,7 @@ class BookController {
   }
   async show(req, res) {
     try {
+
       const { _id } = req.params;
 
       const showBook = await TaskBook.findById(_id);
